dataSource.subscribe checks existance of data not row
Given table binded to an array with 2 items
And table is filtered with XYZ
And none of the 2 items has matching info
When when dataSource updated with new 3 items
And filter XYZ cleared
Then table should clear old data
And show 3 new items

Actual since code checks visible rows and due to filter there are no
visible row, it will not clear table.
To fix it we should check available data.